sae-usage/pyproject.toml
[project]
name = "sae-usage"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "bitsandbytes>=0.48.2",
    "flask>=3.1.2",
    "goodfire>=0.3.5",
    "huggingface-hub>=0.36.0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "safetensors>=0.6.2",
    "scikit-learn>=1.5.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
